Making the converstions section of the code This will be finally making most of the imperal/more american systems to the metric/systems that the most of the worlds uses. This will be 1. F - C 2. Feet - Meter 3. Mile - Kilometer 4. Inch - Cm 5. Acre - Square Meter 6. Pound - Kilogram 7. Gallon - Liter 8. Fluid Oz - Milliters 9. Ounce - Grams 10. Gallons/$ - Litter/$ 11. American Dollar - Country Dollar
adding the currancy from the other countires will be mainly getting the country of orgin, then finding there currancy(using a code or list), then doing the converstion via the american dollor to the countries curancy

just added the code for the miles, km, meters, and feet. that and also the design about how I want the website will be. i havent done anything to the code related to the currancy yet