Currancy conversions this is where we will make the new text file of all the currancy's that i have in my current list of countries. here we will 1. match the country with the correct currancy in the conversion section of the code 2. add either an api or a gernal rule of how much 1 usd per the currancy of the contry selected by the user. 3. The user will see the converstion of both the curency of the country selected to usd and how much 1 liter of gas will cost if its was converted to usd per gallon.    a. and vise versa 4. when the button (Go Back) is pushed, both the country and the currancy will be removed from local stroage.
added more countries, and fixed my code that the if statement for the selected country just changes the exchange rate. that means that I have to rearange the label so i dont have to reapt anything. it even checks if the country is a us territory and also check if said us territory also happens to be pureto rico or palistine (it has no current currency so i just leave it as the american usd beacuse they also take that currency)